+++ sdk/python/packages/flet-map/src/flutter/flet_map/lib/src/polygon_layer.dart
@@ -29,11 +29,7 @@ class PolygonLayerControl extends StatelessWidget with FletStoreMixin {
               "label_text_style", Theme.of(context), const TextStyle())!,
           strokeCap: polygon.getStrokeCap("stroke_cap", StrokeCap.round)!,
           strokeJoin: polygon.getStrokeJoin("stroke_join", StrokeJoin.round)!,
-          points: polygon
-              .get("coordinates", [])!
-              .map((c) => parseLatLng(c))
-              .nonNulls
-              .toList());
+          points: polygon.getLatLngList("coordinates"));
     }).toList();
 
     return BaseControl(