+++ sdk/python/packages/flet-map/src/flutter/flet_map/lib/src/polyline_layer.dart
@@ -23,8 +23,8 @@ class PolylineLayerControl extends StatelessWidget with FletStoreMixin {
           borderColor:
               polyline.getColor("border_color", context, Colors.yellow)!,
           color: polyline.getColor("color", context, Colors.yellow)!,
-          pattern: parseStrokePattern(
-              polyline.get("stroke_pattern"), const StrokePattern.solid())!,
+          pattern: polyline.getStrokePattern(
+              "stroke_pattern", const StrokePattern.solid())!,
           strokeCap: polyline.getStrokeCap("stroke_cap", StrokeCap.round)!,
           strokeJoin: polyline.getStrokeJoin("stroke_join", StrokeJoin.round)!,
           strokeWidth: polyline.getDouble("stroke_width", 1.0)!,
@@ -40,11 +40,7 @@ class PolylineLayerControl extends StatelessWidget with FletStoreMixin {
               .map((e) => parseColor(e, Theme.of(context)))
               .nonNulls
               .toList(),
-          points: polyline
-              .get("coordinates", [])!
-              .map((c) => parseLatLng(c))
-              .nonNulls
-              .toList());
+          points: polyline.getLatLngList("coordinates"));
     }).toList();
 
     return BaseControl(